requirements.txt
bcrypt==5.0.0
packaging==26.0
pyqt-toast-notification==1.3.3
PyQt6==6.10.2
PyQt6-Frameless-Window==0.8.1
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.1
python-dotenv==1.2.2
pywin32==311
QtPy==2.4.3
